+++ app/commands/characters_context/nimble/update_command.rb
@@ -4,7 +4,8 @@ module CharactersContext
   module Nimble
     class UpdateCommand < BaseCommand
       include Deps[
-        cache: 'cache.avatars'
+        cache: 'cache.avatars',
+        add_feat: 'commands.characters_context.nimble.feats.add'
       ]
 
       use_contract do
@@ -39,14 +40,38 @@ module CharactersContext
       def lock_key(input) = "character_update_#{input[:character].id}"
       def lock_time = 0
 
+      def do_prepare(input)
+        input[:skill_points] = input[:character].data.skill_points + 1 if input.key?(:level)
+      end
+
       def do_persist(input)
         input[:character].data = input[:character].data.attributes.merge(input.except(:character, :file, :name).stringify_keys)
         input[:character].assign_attributes(input.slice(:name))
         input[:character].save!
 
+        refresh_feats(input) if input.key?(:level)
         upload_avatar(input)
 
         { result: input[:character] }
+      end
+
+      def refresh_feats(input)
+        feats_relation(input).each do |feat|
+          add_feat.call({ character: input[:character], feat: feat })
+        end
+      end
+
+      def feats_relation(input)
+        data = input[:character].data
+        result =
+          ::Nimble::Feat.where(origin: 1, origin_value: data.main_class).where("conditions ->> 'level' = '#{input[:level]}'")
+        if data.subclass
+          result =
+            result.or(
+              ::Nimble::Feat.where(origin: 2, origin_value: data.subclass).where("conditions ->> 'level' = '#{input[:level]}'")
+            )
+        end
+        result
       end
 
       def upload_avatar(input)